+++ test/unit/specs/Select.spec.js
@@ -1170,6 +1170,28 @@ describe('Select.vue', () => {
       })
     })
 
+    it('should trigger the search:clear event if the search text became empty', (done) => {
+      const vm = new Vue({
+        template: '<div><v-select ref="select" search="foo" @search:clear="foo"></v-select></div>',
+        data: { called: false },
+        methods: {
+          foo (val) {
+            this.called = !this.called
+          }
+        }
+      }).$mount()
+
+      vm.$refs.select.search = 'foo'
+      Vue.nextTick(() => {
+        expect(vm.called).toBe(false)
+        vm.$refs.select.search = ''
+        Vue.nextTick(() => {
+          expect(vm.called).toBe(true)
+          done()
+        })
+      })
+    })
+
     it('can set loading to false from the onSearch callback', (done) => {
       const vm = new Vue({
         template: '<div><v-select loading ref="select" :on-search="foo"></v-select></div>',