+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.4.4
+
+- Added search:clear event.
+- Made scrollbar on the dropdown auto.
+
 ## 2.4.3
 
 - Removed link tag from dropdown list item element.